requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase>=2.8.0
httpx>=0.24,<0.28
websockets>=12.0
pyjwt>=2.8.0
cryptography>=41.0.0

# Development tools
black==23.12.1
mypy==1.8.0
pytest==7.4.3
pytest-asyncio==0.21.1

